quebra a estrutura em 3 principais partes
