+++ Platform.Cli/docker-compose.yml
