+++ Platform.Cli/file_loader.js
@@ -2,6 +2,12 @@ var fs = require('fs');
 var yaml = require('js-yaml');
 module.exports = (function(){
     var self = {};
+    /**
+     * 
+     * @param {String} path path de origem da aplicação de Dominio do Usuario
+     * @description Recupera o nome dos arquivos de dominio do usuario
+     * @returns {Array<String>} Array com o nome dos arquivos que definem o dominio
+     */
     self.getFileNames = function(path){
         self.root = path;
         var filesToLoad = [];
