+++ Platform.Cli/model_builder.js
@@ -1,6 +1,6 @@
 const Sequelize = require('sequelize');
 var Handlebars = require('handlebars');
-
+var root = "../Platform.App/";
 Handlebars.registerHelper( "join", function( obj, sep, options ) {    
     return Object.keys(obj).map(function( item ) {        
         return options.fn( obj[item] );
@@ -28,7 +28,13 @@ module.exports = (function(){
     var hasValue = (array,value) => array.filter(v => v === value).length > 0;
 
     var isRelationship = (value) => ["hasMany","belongsTo","hasAndBelongsTo"].filter(v => v === value).length > 0;
-
+    /**
+     * 
+     * @param {JSON} model é o modelo definido no YAML no formato JSON
+     * @description Este método carrega o modelo do YAML e verifica se ele já foi previamente
+     * carregado e adiciona na estrutra de dados de modelos que serão compilados na aplicação de 
+     * dominio da plataforma
+     */
     self.loadModel = function(model){
         var name = keys(model)[0];
         if (self.model.tables[name]){
@@ -54,19 +60,30 @@ module.exports = (function(){
         });        
     };
 
+    /**
+     * @description Compila o modelo de dados declarados no YAML
+     * para JavaScript/Sequelize através de um template definido 
+     * dentro da pasta Platform.App/node_template/model/domain.tmpl
+     */
     self.compile = ()=>{
         var tables = keys(self.model.tables);
         tables.forEach(t => {
             self.compileTable(self.model.tables[t]);
         });
 
-        var source = fs.readFileSync("node_template/model/domain.tmpl").toString();
+        var source = fs.readFileSync(root+"node_template/model/domain.tmpl").toString();
         var template = Handlebars.compile(source);
         var obj = { "model":self.sequelizeModel, "relations":self.model.relationships};        
         var compiled = template(obj);
         return compiled;
     };
 
+    /**
+     * 
+     * @param {JSON} obj modelo de dominio na estrutura de dados pre compilacao
+     * @param {Array<String>} array lista de atributos do dominio
+     * @description Define o tipo da coluna já no formato do Sequelize
+     */
     var defineType = (obj,array) => array.forEach(v => {
         if (self.seqTypeMap[v]){
             obj.type = self.seqTypeMap[v];
@@ -98,6 +115,12 @@ module.exports = (function(){
     
     self.sequelizeModel = {};
     
+    /**
+     * 
+     * @param {JSON} table estrutura pre compilacao
+     * @description compila um objeto de dominio para ja considerando 
+     * detalhes de chave ou atributos
+     */
     self.compileTable = (table)=>{        
         var tableDefinition = {};              
         table.columns.forEach(c => {
